DistanceAware/v2_soft_label/requirements.txt
torch>=2.0.0
chronos-forecasting
numpy
pandas
matplotlib
seaborn
datasets
huggingface_hub
tqdm
